backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
anthropic==0.42.0
pandas==2.2.2
ta==0.11.0
python-dotenv==1.0.1
httpx==0.27.2
finnhub-python==2.4.20
